refactor(foldable): move asynchronous fold into its own interface
