+++ src/fold/index.ts
@@ -1,1 +1,2 @@
 export { Foldable, Folding } from './foldable';
+export { AsyncFoldable } from './async-foldable';
