+++ src/railway/async-railway.ts
@@ -1,10 +1,9 @@
-import { Folding, Kind, Railway } from './railway';
+import { Railway } from './railway';
 
 type AnyRailway<Acceptable, Unacceptable> = AsyncRailway<Acceptable, Unacceptable> | Railway<Acceptable, Unacceptable>;
 
 export interface AsyncRailway<Acceptable, Unacceptable> {
   andThen<U>(transform: (value: Acceptable) => AnyRailway<U, Unacceptable>): AsyncRailway<U, Unacceptable>;
   orElse<U>(transform: (value: Unacceptable) => AnyRailway<Acceptable, U>): AsyncRailway<Acceptable, U>;
-  fold<T>(folding: Folding<Kind, Acceptable, Unacceptable, T>): Promise<T>;
   withTimeout(ms: number, onTimeout: () => Unacceptable): AsyncRailway<Acceptable, Unacceptable>;
 }
